add check on already notified stream

+++ src/renderer/components/FSService.vue
@@ -33,15 +33,17 @@
           fileWatcher.createWatcher(stream.folderPath, (newFilePath) => {
             if (this.fileIsExist(newFilePath)) return
             let files = File.query().where('streamId', stream.id).orderBy('name').get()
-            files.forEach((file) => {
-              if (!fileHelper.isExist(file.path)) {
-                return File.delete(file.id)
-              } else {
-                if (fileHelper.getCheckSum(file.path) === fileHelper.getCheckSum(newFilePath)) {
-                  return this.updateFile(file.id, newFilePath)
+            if (files && files.length) {
+              files.forEach((file) => {
+                if (!fileHelper.isExist(file.path)) {
+                  return File.delete(file.id)
+                } else {
+                  if (fileHelper.getCheckSum(file.path) === fileHelper.getCheckSum(newFilePath)) {
+                    return this.updateFile(file.id, newFilePath)
+                  }
                 }
-              }
-            })
+              })
+            }
             console.log('New file for uploading', newFilePath)
             const file = this.createFileObject(newFilePath, stream)
             this.insertFile(file)
